refactor permission demo code
