+++ sample/src/main/java/com/thoughtworks/ark/sample/feeds/ui/FeedActivity.kt
@@ -1,16 +1,18 @@
-package com.thoughtworks.ark.sample.permission
+package com.thoughtworks.ark.sample.feeds.ui
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
 import com.thoughtworks.ark.ui.theme.Theme
+import dagger.hilt.android.AndroidEntryPoint
 
-class PermissionSampleActivity : AppCompatActivity() {
+@AndroidEntryPoint
+class FeedActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             Theme {
-                PermissionSampleUI()
+                FeedScreen()
             }
         }
     }
