+++ sample/src/main/java/com/thoughtworks/ark/sample/permission/ui/PermissionActivity.kt
@@ -1,19 +1,16 @@
-package com.thoughtworks.ark.sample.feeds
+package com.thoughtworks.ark.sample.permission.ui
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import com.thoughtworks.ark.sample.feeds.ui.FeedScreen
 import com.thoughtworks.ark.ui.theme.Theme
-import dagger.hilt.android.AndroidEntryPoint
 
-@AndroidEntryPoint
-class FeedDemoActivity : AppCompatActivity() {
+class PermissionActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             Theme {
-                FeedScreen()
+                PermissionSampleScreen()
             }
         }
     }
