+++ sample/src/main/java/com/thoughtworks/ark/sample/permission/ui/ScreenState.kt
@@ -1,0 +1,8 @@
+package com.thoughtworks.ark.sample.permission.ui
+
+internal sealed class ScreenState {
+    object None : ScreenState()
+    object SinglePermissionState : ScreenState()
+    object SinglePermissionWithDialogState : ScreenState()
+    object MultiplePermissionsState : ScreenState()
+}
