+++ sample/src/main/res/values/strings.xml
@@ -4,12 +4,13 @@
     <string name="coroutines_label">Coroutines sample</string>
     <string name="powered_by_label">Powered by</string>
 
+    <string name="action_label_cancel">Not Now</string>
+    <string name="action_label_ok">OK</string>
+
     <string name="theme_color_test_for_compose">theme color test for compose</string>
     <string name="theme_color_test_for_xml">theme color test for xml</string>
     <string name="extended_color_test_for_compose">extended color test for compose</string>
     <string name="extended_color_test_for_xml">extended color test for xml</string>
 
     <string name="request_permission_dialog_content">Hi, Please allow APP to access your location. When location permission is enabled, you can see what\'s new around you.</string>
-    <string name="request_permission_dialog_cancel_button">Not Now</string>
-    <string name="request_permission_dialog_confirm_button">OK</string>
 </resources>
